Fix authentication and runtime issues
- Resolved port conflicts by changing to port 3000
- Updated environment variables to match new port
- Fixed Prisma runtime issues by temporarily disabling adapter
- Added mock authentication for development
- Added missing @auth/prisma-adapter dependency
- Server now runs without errors on localhost:3000
