+++ .gitignore
@@ -52,3 +52,4 @@ temp
 
 
 Sarkarisaathiprototype.md
+login.md
