--- apps/web/auth.ts
@@ -1,62 +1,0 @@
-import NextAuth from 'next-auth'
-import Google from 'next-auth/providers/google'
-import Credentials from 'next-auth/providers/credentials'
-import { PrismaAdapter } from '@auth/prisma-adapter'
-import prisma from '@full-stack-nextjs/db'
-import bcrypt from 'bcryptjs'
-import { z } from 'zod'
-
-export const { handlers, auth, signIn, signOut } = NextAuth({
-  adapter: PrismaAdapter(prisma),
-  providers: [
-    Google({
-      clientId: process.env.GOOGLE_CLIENT_ID!,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
-    }),
-    Credentials({
-      name: 'credentials',
-      credentials: {
-        email: { label: 'Email', type: 'email' },
-        password: { label: 'Password', type: 'password' },
-      },
-      async authorize(credentials) {
-        const parsed = z.object({
-          email: z.string().email(),
-          password: z.string().min(8),
-        }).safeParse(credentials)
-
-        if (!parsed.success) return null
-
-        const user = await prisma.user.findUnique({
-          where: { email: parsed.data.email },
-        })
-
-        if (!user?.password) return null
-
-        const valid = await bcrypt.compare(parsed.data.password, user.password)
-        if (!valid) return null
-
-        return user
-      },
-    }),
-  ],
-  callbacks: {
-    async session({ session, user }) {
-      if (session.user) {
-        session.user.id = user.id
-        session.user.role = (user as any).role
-        session.user.isPremium = (user as any).isPremium
-      }
-      return session
-    },
-    async redirect({ url, baseUrl }) {
-      if (url.startsWith(baseUrl)) return url
-      return baseUrl + '/dashboard'
-    },
-  },
-  pages: {
-    signIn: '/login',
-    error: '/login',
-  },
-  session: { strategy: 'database' },
-})
