+++ apps/web/src/app/(operator)/operator/earnings/page.tsx
