--- apps/web/src/app/dashboard/page.tsx
@@ -1,15 +1,0 @@
-import { redirect } from "next/navigation";
-
-import { auth } from "@/lib/auth";
-
-import Dashboard from "./dashboard";
-
-export default async function DashboardPage() {
-  const session = await auth();
-
-  if (!session?.user) {
-    redirect("/login");
-  }
-
-  return <Dashboard user={session.user} />;
-}
