+++ apps/web/src/auth.config.ts
@@ -1,0 +1,54 @@
+import type { NextAuthConfig } from 'next-auth'
+import Google from 'next-auth/providers/google'
+import Credentials from 'next-auth/providers/credentials'
+
+export const authConfig = {
+  providers: [
+    Google({
+      clientId: process.env.GOOGLE_CLIENT_ID,
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET,
+    }),
+    Credentials({
+      name: 'credentials',
+      credentials: {
+        email: { label: 'Email', type: 'email' },
+        password: { label: 'Password', type: 'password' },
+      },
+      // Note: authorize is only called on the server, not in edge
+      async authorize(credentials) {
+        // This will be overridden in auth.ts with DB logic
+        return null
+      },
+    }),
+  ],
+  pages: {
+    signIn: '/login',
+    error: '/login',
+  },
+  callbacks: {
+    async jwt({ token, user, trigger, session }) {
+      if (user) {
+        token.role = (user as any).role
+        token.id = user.id
+      }
+      return token
+    },
+    async session({ session, token }) {
+      if (token && session.user) {
+        session.user.id = token.id as string
+        session.user.role = token.role as string
+      }
+      return session
+    },
+    authorized({ auth, request: { nextUrl } }) {
+      const isLoggedIn = !!auth?.user
+      const isApiAuthRoute = nextUrl.pathname.startsWith('/api/auth')
+      const isPublicRoute = ['/', '/login', '/register'].includes(nextUrl.pathname)
+
+      if (isApiAuthRoute) return true
+      if (isPublicRoute) return true
+      
+      return isLoggedIn
+    },
+  },
+} satisfies NextAuthConfig
