+++ packages/db/src/index.ts
@@ -1,7 +1,7 @@
 import { env } from "@full-stack-nextjs/env/server";
 import { PrismaNeon } from "@prisma/adapter-neon";
 
-import { PrismaClient } from "../prisma/generated/client";
+import { PrismaClient } from "../prisma/generated";
 
 const adapter = new PrismaNeon({
   connectionString: env.DATABASE_URL,
