+++ pnpm-workspace.yaml
@@ -3,7 +3,7 @@ packages:
   - packages/*
 
 catalog:
-  '@prisma/client': ^7.5.0
+  '@prisma/client': 7.5.0
   '@trpc/client': ^11.13.4
   '@trpc/server': ^11.13.4
   '@vitest/coverage-v8': ^3.1.1
